refactored scheduled jobs
